+++ benchmarks/inference/benchmark_megatron_throughout.py
